+++ README.md
@@ -40,7 +40,7 @@ n
 jason카드: 7클로버, K스페이드
 ```
 
-- [ ] 딜러의 카드 수령 여부를 결정하는 기능
+- [X] 딜러의 카드 수령 여부를 결정하는 기능
 ```markdown
 딜러는 16이하라 한장의 카드를 더 받았습니다.
 ```
