+++ src/main/java/blackjack/domain/Card.java
@@ -1,8 +1,8 @@
-package blackjack;
+package blackjack.domain;
 
 public class Card {
-	private String name;
-	private int value;
+	private final String name;
+	private final int value;
 
 	public Card(String name, int value) {
 		this.name = name;
