+++ src/test/java/blackjack/CardDeckTest.java
@@ -2,6 +2,8 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.*;
 
+import blackjack.domain.Card;
+import blackjack.domain.CardDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
