+++ src/test/java/blackjack/ParticipantTest.java
@@ -2,6 +2,7 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.*;
 
+import blackjack.domain.Participant;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
